mbus: replace outstanding fprintf() errors with mbus_error_str_set()
The mbus-protocol-aux.c module is a problematic mix of library APIs and
application functions.  The library should not log anything to stdout or
stderr but instead leave logging to the application.

This affects uses of the local MBUS_ERROR() macro, which as it has been,
unconditionally uses the fprintf() C function for logging to stderr.  It
is replaced with calls to mbus_error_str_set(), meaning the application
now has to take care of logging in case of error return from libmbus.

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>