mbus: check status of serial line when connecting, handle PTYs as 8N1
When connecting to virtual serial ports, e.g. Linux PTYs, we may not be
able to set the parity bit.  This change adds support for checking the
line parameters and verifying the device is an actual serial port before
attempting to set 8E1, on failure assume PTY and revert to 8N1.

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>