mbus: new function mbus_probe_secondary_range()
This is a replacement to the current mbus_scan_2nd_address_range().  The
new function takes a callback as argument and the old function is just a
wrapper calling the new function with a callback to printf() the result.

The idea is to be able to use mbus_probe_secondary_range() as an actual
library function, where the callback can be used to populate a db.

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>